pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "flask>=3.1.2",
    "lxml>=5.4.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "streamlit>=1.51.0",
    "tabulate>=0.9.0",
    "trafilatura>=2.0.0",
]
